new Compiler class is passed to plugins
all the renderBlock logic moves to a separate class :ok_hand:.
also swap compile() args so files comes first--easier usage if plugin doesn't need compiler.
